Added schema and data for tests in resources files.

+++ server/gdproject/src/main/java/server/gdproject/LandRecord.java
@@ -2,5 +2,5 @@ package server.gdproject;
 
 import org.springframework.data.annotation.Id;
 
-public record LandRecord(String address, String owner, int year, int value, int book, @Id int id) {
+public record LandRecord(String address, String owner, int year, int value, int book, @Id Long id) {
 }
